test Cell pseudo pins
